requirements.txt
streamlit>=1.45.0
Pillow>=12.0.0
numpy>=2.4.0
reportlab>=4.0.0
